Raspi compatibility, SwiftLint

+++ Emulator/LogicBoard/Oscillator.cpp
@@ -46,8 +46,9 @@ Oscillator::nanos()
 #else
     
     struct timespec ts;
+    
     (void)clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec * 1000000000 + ts.tv_nsec;
+    return (u64)ts.tv_sec * 1000000000 + ts.tv_nsec;
     
 #endif
 }
